Rename decrypt mod to database
